% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{addAttributesLogical}
\alias{addAttributesLogical}
\title{Add attributes to a data.tree (numeric)}
\usage{
addAttributesLogical(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame from bugphyzz. Output of \code{prepareData2}.}
}
\value{
A clone of the original data.tree.
}
\description{
\code{addAttributesLogical} adds attributes to a data.tree object.
}
