% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{addAttributesNumeric}
\alias{addAttributesNumeric}
\title{Add attributes to a data.tree object (numeric)}
\usage{
addAttributesNumeric(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame. Output of \code{prepareData2}}
}
\value{
A clone of the data.tree object.
}
\description{
\code{addAttributesNumeric} Add numeric attributes to a data.tree object.
}
