% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{addAttributesRange}
\alias{addAttributesRange}
\title{Add attributes to a data.tree object (range)}
\usage{
addAttributesRange(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame. Output of \code{prepareData2}}
}
\value{
A clone of the input data.tree object with extra attributes.
}
\description{
\code{addAttributesRange} adds range attributes to a data.tree object.
}
