% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addRankPrefix}
\alias{addRankPrefix}
\title{Add rank prefix to NCBI_ID}
\usage{
addRankPrefix(id, rank)
}
\arguments{
\item{id}{A character vector of NCBI IDs (taxids).}

\item{rank}{A character vector of ranks}
}
\value{
A character vector of IDs with prefix.
}
\description{
\code{addRankPrefix} adds a prefix (`'[kpcofgst]__'`) to an NCBI ID (taxid)
}
