% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{annotateInternalNodes}
\alias{annotateInternalNodes}
\title{Annotate internal nodes}
\usage{
annotateInternalNodes(ace, df, phy)
}
\arguments{
\item{ace}{Output of the estimateASR function.}

\item{df}{A data frame.}

\item{phy}{A phylo object.}
}
\value{
A data frame
}
\description{
\code{annotateInternalNodes} annotates the internal nodes of the output of
the `estimateASR` function.
}
