% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{asr}
\alias{asr}
\title{asr}
\usage{
asr(node)
}
\arguments{
\item{node}{A node from a data.tree object.}
}
\value{
None. This functions modifies the data.tree object (R6).
}
\description{
\code{asr} Performs a custom function for ancestral state reconstruction.
It actually is a weighted majority vote for the parent node. This function
is meant to used with transversion functions (Do) in a data.tree object.
}
