% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{asrUpstreamLogical}
\alias{asrUpstreamLogical}
\title{ASR / Upstream}
\usage{
asrUpstreamLogical(node)
}
\arguments{
\item{node}{A node.}
}
\value{
Node/R6. A data.tree object
}
\description{
\code{asrUpstreamLogical} Propagate attributes and scores.
}
