% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{asrUpstreamNumeric}
\alias{asrUpstreamNumeric}
\title{ASR / Upstream (numeric)}
\usage{
asrUpstreamNumeric(node)
}
\arguments{
\item{node}{A node.}
}
\value{
node value.
}
\description{
\code{asrUpstreamNumeric} asr for numeric values.
}
