% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{asrUpstreamRange}
\alias{asrUpstreamRange}
\title{ASR / Upstream (range)}
\usage{
asrUpstreamRange(node)
}
\arguments{
\item{node}{A node.}
}
\value{
A modified node.
}
\description{
\code{asrUpstreamRange} asr for range attributes
}
