% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{calcParentScore}
\alias{calcParentScore}
\title{Calculate parent score}
\usage{
calcParentScore(tbl)
}
\arguments{
\item{tbl}{A data.frame. This data.frame should be stored in a data.tree
node.}
}
\value{
A data.frame
}
\description{
\code{calcParentScore} calculates parent score based on taxa children.
First, the score for each attribute is normalized by taxid. Then,
the score is added for all taxa and re-normalized.
}
