% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataReadyForPropagation.R
\name{calcParentScores}
\alias{calcParentScores}
\title{Calculate parent scores}
\usage{
calcParentScores(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{calcParentScores} calculate parent scores with ASR.
}
