% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{chooseColVal}
\alias{chooseColVal}
\title{Choose column for attribute values}
\usage{
chooseColVal(df)
}
\arguments{
\item{df}{A data frame imported with bugphyzz.}
}
\value{
A character string.
}
\description{
\code{chooseColVal} chooses the column where attribute values are
actually stored.
}
