% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci_to_scores}
\alias{ci_to_scores}
\title{Convert confidence intervals to numeric scores}
\usage{
ci_to_scores(x)
}
\arguments{
\item{x}{A dataset from bugphyzz, e.g., aerophilicity.}
}
\value{
A datafraame. The same dataframe with the addtional `Score` column.
}
\description{
\code{.ci_to_scores} converts the keywords in the `confidence_interval`
column of a bugphyzz dataset into numeric scores, which added as an
additional column named as `Score`.
}
