% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanNode}
\alias{cleanNode}
\title{Clean node}
\usage{
cleanNode(node)
}
\arguments{
\item{node}{A node in a data.tree object.}
}
\value{
NULL (assigned to the node)
}
\description{
\code{cleanNode} deletes the `attribute_tbl` slot in a data.tree node.
}
