% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{completeBinaryData}
\alias{completeBinaryData}
\title{Complete binary data}
\usage{
completeBinaryData(tbl)
}
\arguments{
\item{tbl}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{completeBinaryData} completes Attributes and scores for binary
attributes
}
