% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conf2Fct}
\alias{conf2Fct}
\title{Confidence in curation to factor}
\usage{
conf2Fct(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A factor.
}
\description{
\code{conf2Fct} converts a vector from the Confidence_in_curation
column in a bugphyzz dataset to a factor.
}
