% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{counts_per_attribute_and_rank}
\alias{counts_per_attribute_and_rank}
\title{Counts per attribute and rank}
\usage{
counts_per_attribute_and_rank(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame.
}
\description{
Counts per attribute and rank
}
