% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr-taxonomy.R
\name{createEntriesASR}
\alias{createEntriesASR}
\title{Create entries for ASR}
\usage{
createEntriesASR(x, phy, df)
}
\arguments{
\item{x}{Character states.}

\item{phy}{A phylo object.}

\item{df}{Tree metadata (a data frame).}
}
\value{
A data frame
}
\description{
Create entries for ASR
}
