% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.get_cache}
\alias{.get_cache}
\title{Get NCBI taxids}
\usage{
.get_cache()
}
\arguments{
\item{keyword}{Default 'all'. Options b = base, i = informal,
l = unclassified, t = uncultured.}

\item{full_taxonomy}{Default TRUE.}

\item{force_download}{If FALSE (default) the function uses the ncbi
taxonomy in cache (if present). If TRUE the ncbi taxonomy is dowloaded.
If a taxonomy was present in the cache, it will be removed and replaced.}
}
\value{
A table of taxids with taxonomic information.


get_ncbi_taxonomy

\code{get_ncbi_taxonomy} downloads and imports the full taxonomic
classification from the NCBI database

A data frame with the complete NCBI taxonomy
}
\description{
\code{get_ncbi_taxids} retrieves the taxids from taxPPro. These taxids
correspond to the sets of the taxonomy statistics in the NCBI.
}
