% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downstream}
\alias{downstream}
\title{Get annotations for children node (downstream)}
\usage{
downstream(df)
}
\arguments{
\item{df}{A data frame from bugphyzz}
}
\value{
A data frame with children nodes annotated
}
\description{
Get annotations for children node (downstream)
}
