% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation-asr-custom.R
\name{filter_dataset_for_propagation}
\alias{filter_dataset_for_propagation}
\title{Filter dataset propagation}
\usage{
filter_dataset_for_propagation(df, df_name = NULL)
}
\arguments{
\item{df}{A dataset from bugphyzz.}

\item{df_name}{A character string. The name of the dataset.}
}
\value{
A filtered version of the input dataframe.
If no rows are kept, the output is NULL with a warning.
}
\description{
\code{filter_dataset_for_propagation} filters rows that
have enough information to perform ASR and Inheritance propagation of
taxa annotations.
}
