% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_preprocessing.R
\name{filterData}
\alias{filterData}
\title{Filter data for propagation}
\usage{
filterData(df, df_name = NULL, tax.id.type, remove_false = TRUE)
}
\arguments{
\item{df}{A dataset from bugphyzz.}

\item{df_name}{A character string. The name of the dataset.}

\item{tax.id.type}{A character string. Valid options: NCBI_ID, Taxon_name.
This may be transformed to include other taxonomy IDs.}

\item{remove_false}{If TRUE, row with FALSE attribute values are removed.
Default is FALSE.}
}
\value{
A filtered version of the input data frame.
If no rows are kept, the output is NULL with a warning.
}
\description{
\code{filterData} removes rows that don't have enough information to
perform propagation with ASR and Inheritance.
}
