% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{filterDataDiscrete}
\alias{filterDataDiscrete}
\title{Filter attributes with discrete type}
\usage{
filterDataDiscrete(tbl)
}
\arguments{
\item{tbl}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{filterDataDiscrete} filters discrete data.
}
