% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{filterDataNumeric}
\alias{filterDataNumeric}
\title{Filter data of numeric attributes}
\usage{
filterDataNumeric(tbl)
}
\arguments{
\item{tbl}{A data.frame imported with \code{bugphyzz::physiologies}}
}
\value{
A data.frame.
}
\description{
\code{filterDataNumeric} filters data that could be used for propagation
of numeric attributes. Attribute type is numeric because all numeric
attributes are converted to range when they're imported with the physiologies
function.
}
