% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freq2Scores2}
\alias{freq2Scores2}
\title{Frequency to scores 2}
\usage{
freq2Scores2(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A numeric vector.
}
\description{
\code{freq2Scores} converts the keywords in the "Frequency"
column of a bugphyzz dataset into numeric scores, which are interpreted
as probabilities. This v2 should be used after performing ASR in
numeric attributes.
}
