% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{get_agreements}
\alias{get_agreements}
\title{Get agreements}
\usage{
get_agreements(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{get_agreements} gets taxa annotated two or more times from different
sources.
}
