% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{get_children}
\alias{get_children}
\title{Get children}
\usage{
get_children(x)
}
\arguments{
\item{x}{A vector of valid NCBI taxids.}
}
\value{
A data frame with all children taxa.
}
\description{
\code{get_children} gets the immediate descendants of a taxon.
}
