% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{get_conflicts}
\alias{get_conflicts}
\title{Get conflicts}
\usage{
get_conflicts(df)
}
\arguments{
\item{df}{A data frame from bugphzz}
}
\value{
a data frame or NULL
}
\description{
\code{get_conflicts} gets taxa with two or more annotations from different
sources.
}
