% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{get_double_annotations}
\alias{get_double_annotations}
\title{Get taxa with double annotations}
\usage{
get_double_annotations(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{get_double_annotations} gets taxa annotated twice from the same source.
}
