% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Get duplicates}
\usage{
get_duplicates(df, cols = c("Taxon_name", "NCBI_ID"), verbpse = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{Columns to look for duplicates.}
}
\value{
A dataframe with duplicated rows. If no duplicates are found, this
function returns a NULL value with a message.
}
\description{
\code{get_duplicates} gets duplicated taxa in a bugphuzz dataset.
This information could be useful for identifying conflicts of annotations
between different sources.
}
