% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_ncbi_taxids}
\alias{get_ncbi_taxids}
\title{Get NCBI taxids}
\usage{
get_ncbi_taxids(keyword = "all", full_taxonomy = TRUE)
}
\arguments{
\item{keyword}{Default 'all'. Options b = base, i = informal,
l = unclassified, t = uncultured.}

\item{full_taxonomy}{Default TRUE.}
}
\value{
A table of taxids with taxonomic information.
}
\description{
\code{get_ncbi_taxids} retrieves the taxids from taxPPro. These taxids
correspond to the sets of the taxonomy statistics in the NCBI.
}
