% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parents}
\alias{get_parents}
\title{Get parents}
\usage{
get_parents(x)
}
\arguments{
\item{x}{A vector of valid NCBI IDs}
}
\value{
A table with Parent_NCBI_ID, Parent_name, and Parent_rank
}
\description{
\code{get_parents} Get's the parents of a vector of valid NCBI IDs.
}
