% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{getAgreements}
\alias{getAgreements}
\title{Get agreements}
\usage{
getAgreements(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{getAgreements} gets taxa annotated two or more times with the same
annotation, but from different sources.
}
\details{
An agreement is defined as taxon with the same annotation from different
sources.
}
\seealso{
\code{\link{resolveAgreements}}
}
