% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{getChildren}
\alias{getChildren}
\title{Get children}
\usage{
getChildren(x)
}
\arguments{
\item{x}{A vector of valid NCBI taxids.}
}
\value{
A data frame with all children taxa.
}
\description{
\code{getChildren} gets the immediate descendants of a taxon.
}
