% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{getChildrenScores}
\alias{getChildrenScores}
\title{Get scores for children taxa}
\usage{
getChildrenScores(df)
}
\arguments{
\item{df}{A data frame imported from bugphyzz and output of the `preSteps`
function}
}
\value{
A data frame with scores for children taxa.
}
\description{
\code{getChildrenScores} works
}
