% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{getConflicts}
\alias{getConflicts}
\title{Get conflicts}
\usage{
getConflicts(df)
}
\arguments{
\item{df}{A data frame. Output of the `getDuplicates` function.}
}
\value{
a data frame or NULL.
}
\description{
\code{getConflicts} gets taxa with two or more annotations from different
sources.
}
