% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{getDataReady}
\alias{getDataReady}
\title{Get data ready for propagation}
\usage{
getDataReady(tbl)
}
\arguments{
\item{tbl}{A data.frame.}
}
\value{
A data.frame.
}
\description{
Get data ready for propagation
}
