% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{getDistantTips}
\alias{getDistantTips}
\title{Get distant tips}
\usage{
getDistantTips()
}
\value{
A data.frame.
}
\description{
\code{getDistantTips} gets furthest apart tip for every tip in the ltp tree.
Furthermore, this tip's taxid is present in the ncbi_tree. Additional
columns describe data in the tip2 column.
}
