% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{getDoubleAnnotations}
\alias{getDoubleAnnotations}
\title{Get taxa with double annotations}
\usage{
getDoubleAnnotations(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{getDoubleAnnotations} gets taxa annotated two or more times from the
same source.
}
\seealso{
\code{\link{getDuplicates}}
}
