% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{getMRCATaxPPro}
\alias{getMRCATaxPPro}
\title{Get Most Recent Common Ancestor}
\usage{
getMRCATaxPPro(tree, tips)
}
\arguments{
\item{tree}{A phylo tree object.}

\item{tips}{Tips in the phylo tree object.}
}
\value{
A data.frame.
}
\description{
\code{getMRCATaxPPro} gets the most recent common ancestor using phytools.
}
