% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{getNCBI}
\alias{getNCBI}
\title{Get NCBI data (tree or table)}
\usage{
getNCBI(keyword = "b", format = "table")
}
\arguments{
\item{keyword}{A character string. See}

\item{format}{A character string. Options: 'table' or 'tree'.}
}
\value{
A data.frame or a tree in data.tree format (Node, R6) .
}
\description{
\code{getNCBI} gets data from the NCBI in tree or table format. Tree format
is in the node format of the data.tree package. For now, it only inlcudes
taxa from the bacteria superkingdom/domain/kingdom.
}
