% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{getParents}
\alias{getParents}
\title{Get parents}
\usage{
getParents(x)
}
\arguments{
\item{x}{A vector of valid NCBI taxids}
}
\value{
A data frame with Parent_NCBI_ID, Parent_name, and Parent_rank.
}
\description{
\code{getParents} get's the parent taxon of a vector of valid NCBI IDs.
}
