% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{getParentScores}
\alias{getParentScores}
\title{Calculate parent score}
\usage{
getParentScores(df)
}
\arguments{
\item{df}{A data frame.}

\item{wt}{whether use weights on the dplyr::counts function.}
}
\value{
A data frame.
}
\description{
\code{getParentScores} calculates the parent score based on its immediate
descendants.
}
