% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{getSetWithoutIDs}
\alias{getSetWithoutIDs}
\title{Get set without IDs}
\usage{
getSetWithoutIDs(tbl, set_with_ids = NULL)
}
\arguments{
\item{tbl}{A data.frame.}

\item{set_with_ids}{A data.frame}
}
\value{
A data.frame.
}
\description{
\code{getSetWithoutIDs}
}
