% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_ncbi_ids}
\alias{import_ncbi_ids}
\title{Import NCBI IDs from the statistics page}
\usage{
import_ncbi_ids()
}
\value{
A list of NCBI taxids.
}
\description{
\code{imrpot_ncbi_ids} imports the NCBI ids as grouped in the statistics
page. This data was imported by manually selecting exclusion options in
the NCBI taxonomy statistics page: https://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=statistics&taxon=2157%3AArchaea&taxon=2%3ABacteria&period=&from=&to=
}
\details{
The data was downloaded on
September 2, 2022. Current numbers might be different.
}
