% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{inh}
\alias{inh}
\title{inh}
\usage{
inh(node)
}
\arguments{
\item{node}{A node from a data.tree object.}
}
\value{
None. This functions modifies the data.tree object (R6).
}
\description{
\code{inh} inheritance from parent node to children nodes. This functions is
meant to be used with DO while traversing a data.tree object.
}
