% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{inhDownstreamLogical}
\alias{inhDownstreamLogical}
\title{Inheritance / Downstream (logical)}
\usage{
inhDownstreamLogical(node)
}
\arguments{
\item{node}{A node}
}
\value{
Node, R6, data.tree.
}
\description{
\code{inhDownstreamLogical} each node inherits an attribute from it's
ancestor if values are NA or NULL
}
