% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{inhDownstreamNumeric}
\alias{inhDownstreamNumeric}
\title{Inheritance / Downstream (numeric)}
\usage{
inhDownstreamNumeric(node)
}
\arguments{
\item{node}{A node}
}
\value{
Node, R6, data.tree.
}
\description{
\code{inhDownstreamNumeric} each node inherits an attribute from it's ancestor
if values are NA or NULL
}
