% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{inhDownstreamRange}
\alias{inhDownstreamRange}
\title{Inheritance / Downstream (range)}
\usage{
inhDownstreamRange(node)
}
\arguments{
\item{node}{A node}
}
\value{
Node, R6, data.tree.
}
\description{
\code{inhDownstreamRange} each node inherits an attribute from it's ancestor
if values are NA or NULL
}
