% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp}
\alias{ltp}
\title{Get living tree project (LTP) tree or data}
\usage{
ltp(x = "tree")
}
\arguments{
\item{x}{A character string. tree, tips, or nodes. Default is tree.}
}
\value{
A phylo or data.frame (see the `x` param)
}
\description{
\code{ltp} gets the LTP tree or data.
}
