% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp3}
\alias{ltp3}
\title{Get living tree project (LTP) tree or data v3}
\usage{
ltp3()
}
\value{
A list with a phylo (the ltp tree) and a data.frame (tip_data)
objects.
}
\description{
\code{ltp3} gets the LTP tree or data. Use `ape::drop.tip` to remove
tips.
}
