% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{mergeOriginalAndEarlyASR}
\alias{mergeOriginalAndEarlyASR}
\title{Merge original and early ASR}
\usage{
mergeOriginalAndEarlyASR(l)
}
\arguments{
\item{l}{A list. Output of \code{prepareDataForPropagation}.}
}
\value{
A data.frame.
}
\description{
\code{mergeOriginalAndEarlyASR} merges original bugphyzz annotations
and early ASR entries (those taxids that couldn't be mapped to the
data.tree structures, but their parents could).
}
