% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsti2Freq}
\alias{nsti2Freq}
\title{NSTI to Frequency}
\usage{
nsti2Freq(x)
}
\arguments{
\item{x}{A numeric vectir of nsti values}
}
\value{
A character vector of frequency keywords.
}
\description{
\code{nsti2Freq} converts nsti values to frequencuy keywords.
}
