% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_preprocessing.R
\name{numericToRange}
\alias{numericToRange}
\title{Convert numeric atttributes to range}
\usage{
numericToRange(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data.frame
}
\description{
\code{numericToRange} convers numeric attributes to range attributes.
This function is meant to be used in \code{prepareData2}.
}
