% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_preprocessing.R
\name{prepareData}
\alias{prepareData}
\title{Preparare data for propagation part 1}
\usage{
prepareData(df, tax.id.type = "NCBI_ID", remove_false = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{tax.id.type}{A character string. Valid values are Taxon_name and
NCBI_ID. Default is NCBI_ID.}

\item{remove_false}{If TRUE, Attribute values with FALSE are removed.
Default is FALSE.
NCBI_ID.}
}
\value{
A data.frame.
}
\description{
\code{prepareData} performs the previous steps before propagating
annotations.
These steps include: 1) filter data, 2) resolve conflicts, 3) resolve
agreements, 4) remove duplicates, 5) convert frequency to scores,
6) remove duplicate lines, 7) ensure Taxid and Parent_tax_id are are
character vectors.
}
