% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{prepareData2}
\alias{prepareData2}
\title{Prepare data for propagation part 2}
\usage{
prepareData2(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{prepareData2} prepares data for propagation. Must be used after
\code{prepareData}.
}
