% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepareDataForPropagation}
\alias{prepareDataForPropagation}
\title{Prepare data for propagation}
\usage{
prepareDataForPropagation(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A named list.
}
\description{
\code{prepareDataForPropagation} prepares data for propagation.
}
