% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepareDatForPropagation}
\alias{prepareDatForPropagation}
\title{Prepare Data for Propagation}
\usage{
prepareDatForPropagation(df, resolve = TRUE)
}
\arguments{
\item{df}{A data.frame.}

\item{resolve}{Logical value. If TRUE (deafulat) resolve agreements,
conflicts, and double annotations.}
}
\value{
A data.frame
}
\description{
\code{prepareDataForPropagation} prepares data to be used in the
propagation method.
}
