% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_preprocessing.R
\name{preSteps}
\alias{preSteps}
\title{Steps previous to propagation}
\usage{
preSteps(df, tax.id.type, remove_false = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{tax.id.type}{A character string. Valid values are Taxon_name and}

\item{remove_false}{If TRUE, Attribute values with FALSE are removed.
Default is FALSE.
NCBI_ID.}
}
\value{
A data.frame.
}
\description{
\code{preSteps} performs the previous steps before propagating annotations.
These steps include: 1) filter data, 2) resolve conflicts, 3) resolve
agreements, 4) remove duplicates, 5) convert frequency to scores,
6) remove duplicate lines, 7) ensure Taxid and Parent_tax_id are are
character vectors.
}
