% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printDataTreeAttributes}
\alias{printDataTreeAttributes}
\title{Print data.tree attributes}
\usage{
printDataTreeAttributes(data_tree, limit = 100)
}
\arguments{
\item{data_tree}{A data.tree object with attributes.}

\item{limit}{The number of nodes to be displayed. Default = 100.}
}
\value{
A data.frame or data on the console. Can use \code{View}.
}
\description{
\code{printDataTreeAttributes} prints all of the attributes in a data.tree.
}
