% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propagate}
\alias{propagate}
\title{Propagate annotations}
\usage{
propagate(df)
}
\arguments{
\item{df}{A data frame from bugphyzz.}
}
\value{
A data frame with extended annotations.
}
\description{
Propagate annotations
}
