% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{propagateAnnotations}
\alias{propagateAnnotations}
\title{Propagate annotations}
\usage{
propagateAnnotations(df, max.tax.level, direction)
}
\arguments{
\item{df}{A data frame.}

\item{max.tax.level}{A character string. Maximum taxonomic rank/level.}

\item{direction}{A character string. 'upstream' or 'downstream'.}
}
\value{
A data frame with asr/inh evidence.
}
\description{
\code{propagateAnnotations}
}
