% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{propagateUpstream}
\alias{propagateUpstream}
\title{Propagate upstream}
\usage{
propagateUpstream(df, max.tax.level)
}
\arguments{
\item{df}{A data frame.}

\item{max.tax.level}{A character string. Maximum taxonomic rank/level.}
}
\value{
A data frame with asr evidence.
}
\description{
\code{propagateUpstream}
}
