% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_taxa_duplicates}
\alias{remove_taxa_duplicates}
\title{Remove taxa duplicates
\code{remove_taxa_duplicates} remove taxa that are duplicated}
\usage{
remove_taxa_duplicates(df)
}
\arguments{
\item{df}{A dataframe imported from bugphyzz}
}
\value{
A dataframe without duplicated taxa
}
\description{
Remove taxa duplicates
\code{remove_taxa_duplicates} remove taxa that are duplicated
}
