% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove taxa duplicates
\code{removeDuplicates} removes taxa that are duplicated (output of
the `getDuplicates` function).}
\usage{
removeDuplicates(df)
}
\arguments{
\item{df}{A data frame imported with bugphyzz.}
}
\value{
A data frame without duplicated taxa.
}
\description{
Remove taxa duplicates
\code{removeDuplicates} removes taxa that are duplicated (output of
the `getDuplicates` function).
}
\seealso{
\code{\link{getDuplicates}}
}
