% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{removeTaxaDuplicates}
\alias{removeTaxaDuplicates}
\title{Remove taxa duplicates
\code{removeTaxaDuplicates} removes taxa that are duplicated (output of
the `getDuplicates` function).}
\usage{
removeTaxaDuplicates(df)
}
\arguments{
\item{df}{A data frame imported with bugphyzz.}
}
\value{
A data frame without duplicated taxa.
}
\description{
Remove taxa duplicates
\code{removeTaxaDuplicates} removes taxa that are duplicated (output of
the `getDuplicates` function).
}
