% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{resolve_agreements}
\alias{resolve_agreements}
\title{Resolve conflicts}
\usage{
resolve_agreements(df)
}
\arguments{
\item{df}{A data frame imported from bugphuyzz.}
}
\value{
A data frame
}
\description{
\code{resolve_agreements} resolves agreements, returning only one
(the highest).
}
