% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{Resolve conflicts in a bugphyzz dataset}
\usage{
resolve_conflicts(df)
}
\arguments{
\item{df}{A dataframe imported with bugphyzz}
}
\value{
A dataframe with resolved conflicts
}
\description{
Resolve conflicts in a bugphyzz dataset
}
