% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{resolveAgreements}
\alias{resolveAgreements}
\title{Resolve agreements}
\usage{
resolveAgreements(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{resolveAgreements} resolve all agreements in a bugphyzz dataset.
Agreements happen when the same annotations comes from two or more
sources.
}
