% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundDec}
\alias{roundDec}
\title{Round up to nearest decimal}
\usage{
roundDec(x)
}
\arguments{
\item{A}{character vector (of attributes).}
}
\value{
A double.
}
\description{
\code{roundDec} rounds up to next decimal. Meant to be used for
establishing a threhold for filtering attributes.
}
