% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scores2Freq}
\alias{scores2Freq}
\title{Scores to frequency}
\usage{
scores2Freq(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
\code{scores2Freq} converts numeric scores, stored in the "Score" column in
a bugphyzz dataset, into frequency keywords.
}
