% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{taxname2taxid}
\alias{taxname2taxid}
\title{Conver taxname to taxid}
\usage{
taxname2taxid(tax_tbl)
}
\arguments{
\item{tax_tbl}{An table with taxnames.}
}
\value{
A data.frame.
}
\description{
\code{taxname2taxid} Converts a taxon name from the ncbi to its corresponding
taxid
}
