% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{taxPool}
\alias{taxPool}
\title{Perform taxonomic pooling}
\usage{
taxPool(node, grp, typ)
}
\arguments{
\item{node}{A node in a data.tree object.}

\item{grp}{A character string. Attribute group.}

\item{typ}{A character string. Attribute_type.}
}
\value{
A table for the attribute of a node.
}
\description{
\code{taxPool} performs taxonomic pooling at the strain, species, and
genus ranks. Only use in the data.tree NCBI tree.
}
