% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toDataFrame}
\alias{toDataFrame}
\title{toDataFrame}
\usage{
toDataFrame(data_tree, ncbi_tax)
}
\arguments{
\item{data_tree}{A data.tree.}

\item{ncbi_tax}{NCBI taxonomy. Output of \code{get_ncbi_taxonomy}.}
}
\value{
A data.frame.
}
\description{
\code{toDataFrame}
}
