% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upstream}
\alias{upstream}
\title{Code used to annotate parent nodes according to the NCBI taxonomy}
\usage{
upstream(df)
}
\arguments{
\item{df}{A data frame from bugphyzz}
}
\value{
A datta frame with parent nodes annotations
}
\description{
Code used to annotate parent nodes according to the NCBI taxonomy
}
