% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{addAttributes}
\alias{addAttributes}
\title{Add attributes to a data.tree object}
\usage{
addAttributes(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame. Output of \code{prepareData2}}
}
\value{
A clone of the original data.tree object.
}
\description{
\code{addAttributes} Adds attributes to a data.tree object
}
