% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{addAttributesLogical}
\alias{addAttributesLogical}
\title{Add attributes to data_tree}
\usage{
addAttributesLogical(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame from bugphyzz}
}
\value{
A data.tree with extra attributes.
}
\description{
\code{addAttributesLogical} adds attributes to a data.tree object.
}
