% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcParentScore}
\alias{calcParentScore}
\title{Calculate parent scores}
\usage{
calcParentScore(df)
}
\arguments{
\item{df}{A data frame.}

\item{wt}{If use weights on the dplyr::counts function}
}
\value{
A data frame.
}
\description{
\code{calcParentScore} calculates the parent scores in a bugphyzz dataset.
}
