% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation-asr-custom.R
\name{calcParentScore}
\alias{calcParentScore}
\title{Calculate parent score}
\usage{
calcParentScore(df)
}
\arguments{
\item{df}{A data frame.}

\item{wt}{whether use weights on the dplyr::counts function}
}
\value{
A data frame.
}
\description{
\code{calcParentScore} calculates the parent score based on its immediate
descendants.
}
