% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{downstream}
\alias{downstream}
\title{Get annotations for descendants (downstream)}
\usage{
downstream(df)
}
\arguments{
\item{df}{A data frame from bugphyzz.}
}
\value{
A data frame with child nodes annotated (new rows).
}
\description{
\code{downstream} annotates the descendants fo parent nodes.
}
