% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_dataset_for_propagation}
\alias{filter_dataset_for_propagation}
\title{Filter datset for phylogenetic propagation}
\usage{
filter_dataset_for_propagation(df, df_name = NULL)
}
\arguments{
\item{df}{A dataset from bugphyzz.}

\item{df_name}{A character string. The name of the dataset}
}
\value{
A flitered version of the input df. If no rows are conserved, the
output is NULL with a warning.
}
\description{
\code{filter_dataset_for_propagation} filters rows and datasets that
have enough information to perfrom ASR and Inheritance propagation of
taxa annotations
}
