% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{filterData}
\alias{filterData}
\title{Filter data}
\usage{
filterData(tbl)
}
\arguments{
\item{tbl}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{filterData} is the first step in the propagation process.
It ensures that only data that can be used for propagation is retained.
}
