% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{filterDataNumeric}
\alias{filterDataNumeric}
\title{Filter data with numeric attributes}
\usage{
filterDataNumeric(tbl)
}
\arguments{
\item{tbl}{A data.frame imported with \code{bugphyzz::physiologies}}
}
\value{
A data.frame.
}
\description{
\code{filterDataNumeric} filters data that could be used for propagation
of numeric attributes. These are labeled as "range" when being imported
with the \code{physiologies function}. It also creates a Score column
based on the Frequency column.
}
