% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_children}
\alias{get_children}
\title{Get children taxa}
\usage{
get_children(x)
}
\arguments{
\item{x}{A vector of NCBI IDs}
}
\value{
A dataframe with all children taxa
}
\description{
Get children taxa
}
