% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation-asr-custom.R
\name{get_children}
\alias{get_children}
\title{Get children}
\usage{
get_children(x)
}
\arguments{
\item{x}{A vector of NCBI IDs}
}
\value{
A data frame with all children taxa
}
\description{
\code{get_children} gets the immediate descendants of a taxon.
}
