% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{getAgreements}
\alias{getAgreements}
\title{Get agreements}
\usage{
getAgreements(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{getAgreements} gets taxa annotated two or more times from different
sources.
}
