% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_ncbi_ids}
\alias{import_ncbi_ids}
\title{Import NCBI IDs from the statistics page}
\usage{
import_ncbi_ids()
}
\value{
A list of NCBI taxids.
}
\description{
\code{imrpot_ncbi_ids} imports the NCBI ids as grouped in the statistics
page.
}
