% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPPro.R
\name{inh1}
\alias{inh1}
\title{Inheritance first round}
\usage{
inh1(node, adjF = 0.1)
}
\arguments{
\item{node}{A node in a data.tree.}

\item{adjF}{Adjustment factor for penalty. Default is 1.0.}
}
\value{
A table for a node attribute.
}
\description{
\code{inh1} First round of inheritance. Only use with Do in a data.tree
object.
}
