% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp}
\alias{ltp}
\title{Get living tree project (LTP) tree or data}
\usage{
ltp(data = FALSE)
}
\arguments{
\item{data}{TRUE for data. FALSE for tree. Default is FALSE.}
}
\value{
A phylo or data.frame (see the `data` param)
}
\description{
\code{ltp} gets the LTP tree or data.
}
