% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp}
\alias{ltp}
\title{Get living tree project (LTP) tree and data}
\usage{
ltp(remove_gn_nodes = TRUE, node_names = TRUE)
}
\arguments{
\item{remove_gn_nodes}{Remove rows from node_data with genus taxids.
Default is TRUE, which removes them. These are already in tip_data.}

\item{node_names}{Add names to unnamed nodes (n + node number).
This will only be in the tree (not in node_data).}
}
\value{
A list with a the LTP tree (phylo), tips and node data (data.frames),
and the names of the genera added to the original tree (character vector).
.
}
\description{
\code{ltp} Imports a modified version fo the phylogenetic tree of the
Living Tree Project.
}
