% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp}
\alias{ltp}
\title{Get living tree project (LTP) tree and data}
\usage{
ltp()
}
\value{
A list with a the LTP tree (phylo), tips and node data (data.frames),
and the names of the genera added to the original tree (character vector).
.
}
\description{
\code{ltp} Imports a modified version fo the phylogenetic tree of the
Living Tree Project.
}
