% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{ltp}
\alias{ltp}
\title{Get living tree project (LTP) tree or data}
\usage{
ltp(remove_zero_tips = TRUE)
}
\arguments{
\item{remove_zero_tips}{Remove or not tips with zero branch lengths.
These tips could cause trouble.}
}
\value{
A list with a phylo (the ltp tree) and a data.frame (tip_data)
objects.
}
\description{
\code{ltp} gets the LTP tree or data.
}
