% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{mpa}
\alias{mpa}
\title{Get metphlan tree or data v31}
\usage{
mpa(x = "tree")
}
\arguments{
\item{x}{A character string: tree, tips, or nodes. Default is tree.}
}
\value{
A phylo or data.frame (see the `x` param)
}
\description{
\code{mpa} gets the metaphlan tree or data v31.
}
