% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{mpa}
\alias{mpa}
\title{Get metphlan tree or data v31}
\usage{
mpa(data = FALSE)
}
\arguments{
\item{data}{TRUE for data. FALSE for tree. Default is FALSE.}
}
\value{
A phylo or data.frame (see the `data` param)
}
\description{
\code{mpa} gets the metaphlan tree or data v31
}
