% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataReadyForPropagation.R
\name{prepareDatForPropagation}
\alias{prepareDatForPropagation}
\title{Prepare Data for Propagation}
\usage{
prepareDatForPropagation(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame
}
\description{
\code{prepareDataForPropagation} prepares data to be used in the
propagation method.
}
