% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{propagate}
\alias{propagate}
\title{Propagate annotations}
\usage{
propagate(df, max.tax.level = "genus", direction = "both")
}
\arguments{
\item{df}{A data frame from bugphyzz.}

\item{max.tax.level}{A character string. The method that should be used for
propagation.}

\item{direction}{A character string. 'upstream', 'downstream', or 'both'.}
}
\value{
A data frame with propagated annotations.
}
\description{
\code{propagate} propagates annotations.
}
\examples{

library(bugphyzz)
aer <- physiologies('aerophilicity')[[1]]
aer_prop <- propagate(aer, max.tax.level = 'genus', direction = 'both')

}
