% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{propagate}
\alias{propagate}
\title{propagate}
\usage{
propagate(data_tree, df)
}
\arguments{
\item{data_tree}{A data.tree object.}

\item{df}{A data.frame.}
}
\value{
A data.tree
}
\description{
\code{propagate} propagates asr/inheritance.
}
