% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{propagate}
\alias{propagate}
\title{Propagate annotations}
\usage{
propagate(
  df,
  max.tax.level = "genus",
  direction = "both",
  remove_false = TRUE,
  filter.tax.id.type = "Taxon_name"
)
}
\arguments{
\item{df}{A data frame from bugphyzz.}

\item{max.tax.level}{A character string. The method that should be used for
propagation.}

\item{direction}{A character string. 'upstream', 'downstream', or 'both'.}

\item{remove_false}{If TRUE,remove rows with Attribute value FALSE. Default
is TRUE.}

\item{filter.tax.id.type}{Choose type of id/name to filter. Options are:
'NCBI_ID' and 'Taxon_name'.}
}
\value{
A data frame with propagated annotations.
}
\description{
\code{propagate} propagates annotations.
}
\examples{

library(bugphyzz)
aer <- physiologies('aerophilicity')[[1]]
aer_prop <- propagate(
    aer, max.tax.level = 'genus', direction = 'both', remove_false = TRUE,
    filter.tax.id.type = 'Taxon_name'
)

}
