% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation-asr-custom.R
\name{propagate}
\alias{propagate}
\title{Propagate annotations based on custom ASR}
\usage{
propagate(df, method = "mv")
}
\arguments{
\item{df}{A data frame from bugphyzz.}
}
\value{
A data frame with extended annotations.
}
\description{
\code{propagate} propagates annotations upstrean and downstream the
taxonomy classification of the taxa.
}
