% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation-asr-custom.R
\name{propagate}
\alias{propagate}
\title{Propagate annotations}
\usage{
propagate(df, asr_method = "mv")
}
\arguments{
\item{df}{A data frame from bugphyzz.}

\item{asr_method}{A character string. The method that should be used for
propagation. Option: mv, majority vote; tx, NCBI taxonomy; ph, phylogenetic
tree.}
}
\value{
A data frame with extended annotations.
}
\description{
\code{propagate} propagates annotations.
}
