% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_duplicates.R
\name{resolveAgreements}
\alias{resolveAgreements}
\title{Resolve agreements}
\usage{
resolveAgreements(df)
}
\arguments{
\item{df}{A data frame imported from bugphuyzz.}
}
\value{
A data frame
}
\description{
\code{resolveAgreements} resolves agreements.
}
\details{
Agreements are defined as a taxon with the same annotation from two or more
sources. The agreements are resolved with
`dplyr::slice_max(x, with_ties = FALSE)`. So, only one source is kept,
the one with the highest 'confidence in curation' value.
}
\seealso{
\code{\link{getAgreements}}
}
