% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{resolveAgreements}
\alias{resolveAgreements}
\title{Resolve conflicts}
\usage{
resolveAgreements(df)
}
\arguments{
\item{df}{A data frame imported from bugphuyzz.}
}
\value{
A data frame
}
\description{
\code{resolveAgreements} resolves agreements, returning only one
(the highest).
}
