% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{resolveConflicts}
\alias{resolveConflicts}
\title{Resolve conflicts}
\usage{
resolveConflicts(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{resolveConflicts} resolves are conflicting annotations in a bugphyzz
dataset. Conflicts happen when a single taxon recieves two or more
annotations from different sources.
}
