% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002.duplicates.R
\name{resolveConflicts}
\alias{resolveConflicts}
\title{Resolve conflicts in a bugphyzz dataset}
\usage{
resolveConflicts(df)
}
\arguments{
\item{df}{A dataframe imported with bugphyzz}
}
\value{
A dataframe with resolved conflicts
}
\description{
\code{resolveConflicts} resolves conflicts of categorical values.
}
\details{
A conflict is defined as a taxon with two different annotations from two
different attribute sources.

The conflicts are resolved by giving preference to the source with higher
'confidence in curation' value. This is done with `dplyr::slice_max`. Ties
are resolved according to the output of `dplyr::slice_max`.
}
