% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scores2Freq}
\alias{scores2Freq}
\title{Scores to frequency}
\usage{
scores2Freq(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector
}
\description{
\code{scores2Freq} performs the opposite of \code{freq2scores}.
}
