% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upstream}
\alias{upstream}
\title{Upstream}
\usage{
upstream(df, rank1, rank2)
}
\arguments{
\item{df}{A dataframe from bugphyzz.}

\item{rank1}{Source rank. A character string.}

\item{rank2}{Target rank. A character string}
}
\value{
A dataframe
}
\description{
\code{upstream} propagates annotations upstream in the taxonomy
classification according to the NCBI.
}
