% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_propagation_majority_vote.R
\name{upstream}
\alias{upstream}
\title{Annotate parent nodes (upstream)}
\usage{
upstream(df)
}
\arguments{
\item{df}{A data frame from bugphyzz.}
}
\value{
A data frame with parent nodes annotations (new rows).
}
\description{
\code{upstream} annotates parent node based on their immediate descendants
(children).
}
